% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcanalysis.R
\name{pca.plot}
\alias{pca.plot}
\title{Principal Component Analysis Plotting function}
\usage{
pca.plot(pca)
}
\arguments{
\item{pca}{principal component object returned from the run.pca() function}
}
\description{
This function outputs biplots for principal components 1 vs. 2 and 1 vs. 3, and plots the percentage of variance explained by each component.
}
\examples{
pca.plot(data.pca)
}
\keyword{pca}
\keyword{plot}
\keyword{plotting}
