% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcanalysis.R
\name{pca.scree}
\alias{pca.scree}
\title{Screeplot for k-means analysis}
\usage{
pca.scree(data, kmax, iter.max = 10)
}
\arguments{
\item{data}{input data matrix or data frame (numerical)}

\item{kmax}{maximum k value (numerical)}

\item{iter.max}{maximum number of iterations for k-means (numerical, default is 10)}
}
\description{
This function performs a k-means analysis for each k value and outputs a scree plot of the total within sum of squares for cross-validation.
}
\examples{
pca.scree(data, 10, iter.max = 10)
}
\keyword{kmeans}
\keyword{screeplot}
