% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcanalysis-package.R
\docType{package}
\name{pcanalysis-package}
\alias{pcanalysis}
\alias{pcanalysis-package}
\title{pcanalysis: Principal Components Analysis Package}
\description{
This package can perform a principal component analysis, output biplots, and cluster the data using k-means analysis. Requires a data matrix as input.
}
\keyword{internal}
